+++ relay-server/src/services/processor/span/processing.rs
@@ -413,13 +413,20 @@ fn get_normalize_span_config<'a>(
 
 fn set_segment_attributes(span: &mut Annotated<Span>) {
     let Some(span) = span.value_mut() else { return };
+    let Some(span_id) = span.span_id.value() else {
+        return;
+    };
 
-    // TODO: A span might be a segment span even if the parent_id is not empty
-    // (parent within a trace). I.e. do not overwrite here.
-    let is_segment = span.parent_span_id.is_empty();
+    if let Some(segment_id) = span.segment_id.value() {
+        // The span is a segment if and only if the segment_id matches the span_id.
+        span.is_segment = (segment_id == span_id).into();
+    } else if span.parent_span_id.is_empty() {
+        // If the span has no parent, it is automatically a segment:
+        span.is_segment = true.into();
+    }
 
-    span.is_segment = Annotated::new(is_segment);
-    if is_segment {
+    // If the span is a segment, always set the segment_id to the current span_id:
+    if span.is_segment.value() == Some(&true) {
         span.segment_id = span.span_id.clone();
     }
 }
@@ -638,6 +645,7 @@ mod tests {
     use relay_event_schema::protocol::{
         Context, ContextInner, SpanId, Timestamp, TraceContext, TraceId,
     };
+    use relay_protocol::get_value;
     use relay_sampling::evaluation::{ReservoirCounters, ReservoirEvaluator};
     use relay_system::Addr;
 
@@ -754,5 +762,76 @@ mod tests {
             "{:?}",
             state.envelope()
         );
+    }
+
+    #[test]
+    fn segment_no_overwrite() {
+        let mut span: Annotated<Span> = Annotated::from_json(
+            r#"{
+            "is_segment": true,
+            "span_id": "fa90fdead5f74052",
+            "parent_span_id": "fa90fdead5f74051"
+        }"#,
+        )
+        .unwrap();
+        set_segment_attributes(&mut span);
+        assert_eq!(get_value!(span.is_segment!), &true);
+        assert_eq!(get_value!(span.segment_id!).0.as_str(), "fa90fdead5f74052");
+    }
+
+    #[test]
+    fn segment_overwrite_because_of_segment_id() {
+        let mut span: Annotated<Span> = Annotated::from_json(
+            r#"{
+         "is_segment": false,
+         "span_id": "fa90fdead5f74052",
+         "segment_id": "fa90fdead5f74052",
+         "parent_span_id": "fa90fdead5f74051"
+     }"#,
+        )
+        .unwrap();
+        set_segment_attributes(&mut span);
+        assert_eq!(get_value!(span.is_segment!), &true);
+    }
+
+    #[test]
+    fn segment_overwrite_because_of_missing_parent() {
+        let mut span: Annotated<Span> = Annotated::from_json(
+            r#"{
+         "is_segment": false,
+         "span_id": "fa90fdead5f74052"
+     }"#,
+        )
+        .unwrap();
+        set_segment_attributes(&mut span);
+        assert_eq!(get_value!(span.is_segment!), &true);
+        assert_eq!(get_value!(span.segment_id!).0.as_str(), "fa90fdead5f74052");
+    }
+
+    #[test]
+    fn segment_no_parent_but_segment() {
+        let mut span: Annotated<Span> = Annotated::from_json(
+            r#"{
+         "span_id": "fa90fdead5f74052",
+         "segment_id": "ea90fdead5f74051"
+     }"#,
+        )
+        .unwrap();
+        set_segment_attributes(&mut span);
+        assert_eq!(get_value!(span.is_segment!), &false);
+        assert_eq!(get_value!(span.segment_id!).0.as_str(), "ea90fdead5f74051");
+    }
+
+    #[test]
+    fn segment_only_parent() {
+        let mut span: Annotated<Span> = Annotated::from_json(
+            r#"{
+         "parent_span_id": "fa90fdead5f74051"
+     }"#,
+        )
+        .unwrap();
+        set_segment_attributes(&mut span);
+        assert_eq!(get_value!(span.is_segment), None);
+        assert_eq!(get_value!(span.segment_id), None);
     }
 }